ppv-platform/django_app/requirements.txt
Django>=4.2
psycopg2-binary
redis
stripe
python-dotenv
djangorestframework
